+++ packages/server/src/controller/admin/tag/tag.controller.ts
@@ -8,6 +8,7 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+import { config } from 'src/config';
 import { AdminGuard } from 'src/provider/auth/auth.guard';
 import { ISRProvider } from 'src/provider/isr/isr.provider';
 import { TagProvider } from 'src/provider/tag/tag.provider';
@@ -42,6 +43,12 @@ export class TagController {
     @Param('name') name: string,
     @Query('value') newName: string,
   ) {
+    if (config.demo && config.demo == 'true') {
+      return {
+        statusCode: 401,
+        message: '演示站禁止修改此项！',
+      };
+    }
     const data = await this.tagProvider.updateTagByName(name, newName);
     this.isrProvider.activeAll('批量更新标签名触发增量渲染！');
     return {
@@ -51,6 +58,12 @@ export class TagController {
   }
   @Delete('/:name')
   async deleteTagByName(@Param('name') name: string) {
+    if (config.demo && config.demo == 'true') {
+      return {
+        statusCode: 401,
+        message: '演示站禁止修改此项！',
+      };
+    }
     const data = await this.tagProvider.deleteOne(name);
     this.isrProvider.activeAll('批量删除标签触发增量渲染！');
     return {
